Moved WorldGlobalActions into its own file
